rng added for #13 to make reliable seeding
