+++ Cargo.toml
@@ -4,3 +4,5 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+rand = "0.8"
+rand_chacha = "0.8"
