+++ src/main.rs
@@ -1,5 +1,9 @@
-mod map_gen;
+pub mod models;
+pub mod map_gen;
+#[allow(dead_code)]
+use crate::map_gen::generate_map;
+use crate::models::TileType;
 
 fn main() {
-    println!("Hello, world!");
+ println!("Hello World!");
 }
